api/requirements.txt
annotated-types
anyio
boto3
botocore
click
exceptiongroup
fastapi
h11
idna
jmespath
pillow
pydantic
pydantic_core
pypng
python-dateutil
python-dotenv
qrcode
s3transfer
six
sniffio
starlette
typing_extensions
urllib3
uvicorn
